# fix: Restrict intermediate queries to population using JOIN