Previously we ran every intermediate query against all the patients that happened to be in the table being queried and then only filtered the results down to the defined population when combining the results at the end.

This has the advantage of simplicity but it does mean the intermediate results tables are larger than they need to be and, potentially, the intermediate queries are slower.

This PR changes the logic so we first build a table containing just the patient IDs we want in our population and then use this as the left-most table in any subsequent joins.

There will be contexts in which this results in faster queries and contexts in which it results in slower ones. Our current hypothesis is that it will be a net performance benefit and that the worst case slowdowns will not be pathalogical. But until we test this on some realistic dataset definitions we won't know.

Note that the current implementation is not particularly lovely as it makes use of instance state to avoid passing the population table around, and thus necessitates some cleanup at the end. If the approach turns out to be a win then we can consider a cleaner factoring.

For interest, here is the original year-old Slack thread in which this approach was discussed:
https://bennettoxford.slack.com/archives/C03FB777L1M/p1657899866106729?thread_ts=1657897715.091099&cid=C03FB777L1M