@Jongmassey 2023-08-07
How would this behave with a population definition that references a computed column (e.g. _only include patients who have more than 5 prescriptions of this medication in the past year_ )? 

As I read it, the computed column that supports the population would be included in the `WHERE` clause of the `population_query` but not the `SELECT` list and would then be computed again later in a per-variable temp-table (assuming it was itself included as a variable).