@evansd 2023-08-08
Queries that form part of the population definition won't benefit at all from this, they'll just get compiled exactly as they would before. But there shouldn't be any duplication: any temp tables created as part of the population query will get re-used if they're referenced again later.