@Jongmassey 2023-08-08
>Queries that form part of the population definition won't benefit at all from this, 

Yes I guess that's perfectly logical.

> any temp tables created as part of the population query will get re-used if they're referenced again later

that's what I was worried about, brill