@evansd 2023-08-11
I just tried re-running [this job](https://jobs.opensafely.org/the-healthcare-utilisation-of-people-with-long-covid-openprompt/healthcare_utilisation_openprompt/19635/k3tlmdgtryog65j7/) using the code on this branch. This is the kind of job for which we'd expect the most improvement because it has a tiny, easily defined population so it's in no sense representative but it's good to see what the best case looks like.

It's _quite_ good: total runtime dropped from 2 hours to 6 minutes.