@evansd 2023-08-14
For the less happy path I chose [this job](https://jobs.opensafely.org/impact-of-covid-19-on-diabetes-care-in-england/covid-diabetes-care/19560/sbweasdv44dbmtt6/) which has a population definition of:
```py
dataset.define_population(
    (dataset.age < 110) & current_registration.exists_for_patient()
)
```

Resulting in a population of just under 26 million, which is generally as large as we get.

Previously this took 34 minutes to get to the point of downloading results; using the join on the population table it took 4 hours 30 minutes.

In particular, a couple of queries jumped from taking 3.5 minutes to taking 1 hour 15 minutes (query 005/011 and query 007/013 in the logs below). Both of these were queries on the `CodedEvent` table which had changed from (effectively):
```sql
SELECT ... FROM CodedEvent
```
To
```sql
SELECT ... FROM population_table
JOIN CodedEvent ON population_table.patient_id = CodedEvent.patient_id
```

Logs:
 * [query-timing-log-original.png](https://github.com/opensafely-core/ehrql/assets/19630/2940e526-bb63-4344-a254-b72bc2209cb7)
 * [query-timing-log-with-join.png](https://github.com/opensafely-core/ehrql/assets/19630/f8027d05-214e-4b3f-9c68-af6075b9e69c)
 * [query-original.sql.txt](https://github.com/opensafely-core/ehrql/files/12333690/query-original.sql.txt)
 * [query-with-join.sql.txt](https://github.com/opensafely-core/ehrql/files/12333694/query-with-join.sql.txt)