@evansd 2023-08-14
Disappointed by the above, I tried a new approach which was, instead of a join, to add a WHERE clause like `patient_id IN (SELECT patient_id from #population_table)` to the relevant queries:

```diff
     def get_select_query_for_node_domain(self, node):
         """
         Given a many-rows-per-patient node, return the SELECT query corresponding to
         domain of that node
         """
         frame = get_domain(node).get_node()
         select_table, conditions = get_table_and_filter_conditions(frame)
         table = self.get_table(select_table)
         where_clauses = [self.get_predicate(condition) for condition in conditions]
+        if self.population_table is not None:
+            where_clauses.append(
+                table.c.patient_id.in_(
+                    sqlalchemy.select(self.population_table.c.patient_id)
+                )
+            )
         if where_clauses:
             query = query.where(sqlalchemy.and_(*where_clauses))
         return query
```

This is much more promising:

Job | Population | Original Runtime | With `JOIN` | With `IN (SELECT)`
-- | -- | -- | -- | --
[OpenPrompt](https://jobs.opensafely.org/the-healthcare-utilisation-of-people-with-long-covid-openprompt/healthcare_utilisation_openprompt/19635/k3tlmdgtryog65j7/) | 1,200 | 2 hours | 6 minutes | 8 minutes
[Diabetes](https://jobs.opensafely.org/impact-of-covid-19-on-diabetes-care-in-england/covid-diabetes-care/19560/sbweasdv44dbmtt6/) | 26 million | 30 minutes | 4.5 hours | 43 minutes

I wouldn't pay too much attention to the precise timings here as the server was at different levels of busyness and the OpenPrompt population increased between runs.

But it looks like this approach to filtering might give us some very significant speed-ups in the best case while not costing us too much in the worst case.