@Jongmassey 2023-08-14
How about an `exists` rather than an `in` ? (sometimes can be better)