@evansd 2023-08-14
Because Someone On the Internet had a post showing identical query plans for the `IN` and `EXISTS` versions and the `IN` was syntactically simpler. I can well believe they might have different plans under some circumstances though.