@evansd 2023-08-14
Closing in favour of:
 * https://github.com/opensafely-core/ehrql/pull/1490